+++ R/getLV.gllvm.R
@@ -2,7 +2,7 @@
 #' @description  Extract latent variables from gllvm object.
 #' 
 #' @param object an object of class 'gllvm'.
-#' @param type type of latent variable scores to retrieve from a gllvm object. For models with unconstrained latent variables, defaults to "unconstrained". For models with constrained latent variables, defaults to "constrained". A third option is "scaled", which returns latent variables multiplied with their standard deviations. Alternatively, "LC" returns linear combination scores without residual error.
+#' @param type type of latent variable scores to retrieve from a gllvm object. For models with unconstrained latent variables, defaults to "unconstrained". For models with constrained latent variables, defaults to "constrained". A third option is "scaled", which returns latent variables multiplied with scale parameter of the loadings. Alternatively, "LC" returns linear combination scores without residual error.
 #' @param ... not used
 #' 
 #'@aliases getLV getLV.gllvm