+++ R/gllvm.auxiliary.R
@@ -2064,13 +2064,13 @@ sdrandom<-function(obj, Vtheta, incl, ignore.u = FALSE,return.covb = FALSE){
     
     if(num.RR>0){
       row.names(A)[row.names(A)==""] <- rep("XB",num.RR*n)
-      if(num.lv.c>0){
-        sigma.lv <- c(sigma.lv[1:num.lv.c],rep(1,num.RR),sigma.lv[-c(1:num.lv)])
-      }else if(num.lv>0){
-        sigma.lv <- c(rep(1,num.RR),sigma.lv)
-      }else{
-        sigma.lv <- rep(1,num.RR)
-      }
+      # if(num.lv.c>0){
+      #   sigma.lv <- c(sigma.lv[1:num.lv.c],rep(1,num.RR),sigma.lv[-c(1:num.lv)])
+      # }else if(num.lv>0){
+      #   sigma.lv <- c(rep(1,num.RR),sigma.lv)
+      # }else{
+      #   sigma.lv <- rep(1,num.RR)
+      # }
     }
 
     #Matrix Q for predictors
@@ -2078,7 +2078,7 @@ sdrandom<-function(obj, Vtheta, incl, ignore.u = FALSE,return.covb = FALSE){
     if((num.lv.c+num.lv+radidx)==0)A<-Q
     if((num.lv.c+num.RR)>0){
       for(q in 1:(num.lv.c+num.RR)){
-        Q[(1:n)+n*(q-1)+radidx,which(names(obj$par[incl])=="b_lv")[(1:ncol(lv.X))+(ncol(lv.X)*(q-1))]] <- lv.X/sigma.lv[q]
+        Q[(1:n)+n*(q-1)+radidx,which(names(obj$par[incl])=="b_lv")[(1:ncol(lv.X))+(ncol(lv.X)*(q-1))]] <- lv.X#/sigma.lv[q]
       }
     }
     diag.term2 <- (Q+A)%*%Vtheta%*%t(Q+A)
@@ -2299,20 +2299,20 @@ CMSEPf <- function(fit, return.covb = F){
     if((num.lv+num.lv.c+radidx)==0){colnames(D)<-rep("",ncol(D))}
     colnames(D)[colnames(D)==""]<-"XB"
     row.names(D)<-colnames(D)
-    if(num.lv.c>0){
-      fit$params$sigma.lv <- c(fit$params$sigma.lv[1:num.lv.c],rep(1,num.RR),fit$params$sigma.lv[-c(1:num.lv)])
-    }else if(num.lv>0){
-      fit$params$sigma.lv <- c(rep(1,num.RR),fit$params$sigma.lv)
-    }else{
-      fit$params$sigma.lv <- rep(1,num.RR)
-    }
+    # if(num.lv.c>0){
+    #   fit$params$sigma.lv <- c(fit$params$sigma.lv[1:num.lv.c],rep(1,num.RR),fit$params$sigma.lv[-c(1:num.lv)])
+    # }else if(num.lv>0){
+    #   fit$params$sigma.lv <- c(rep(1,num.RR),fit$params$sigma.lv)
+    # }else{
+    #   fit$params$sigma.lv <- rep(1,num.RR)
+    # }
   }
 
   Q <- matrix(0,nrow=(num.lv+num.lv.c+num.RR)*n+radidx,ncol=dim(A)[1])
 
   if((num.lv.c+num.RR)>0){
     for(q in 1:(num.lv.c+num.RR)){
-      Q[(1:n)+n*(q-1)+radidx,which(names(fit$TMBfn$par[fit$Hess$incl])=="b_lv")[(1:ncol(fit$lv.X))+(ncol(fit$lv.X)*(q-1))]] <- fit$lv.X/fit$params$sigma.lv[q]
+      Q[(1:n)+n*(q-1)+radidx,which(names(fit$TMBfn$par[fit$Hess$incl])=="b_lv")[(1:ncol(fit$lv.X))+(ncol(fit$lv.X)*(q-1))]] <- fit$lv.X#/fit$params$sigma.lv[q]
     }
     }
     covb <- (Q+D%*%C)%*%(A)%*%(t(Q)+B%*%t(D))