+++ R/randomCoefplot.R
@@ -48,7 +48,7 @@ randomCoefplot.gllvm <- function(object, y.label = TRUE, cex.ylab = 0.5, mfrow =
     m <- length(labely)
     Xc <- Xcoef
     
-    if(object$method == "VA"){
+    if((object$method %in% c("VA", "EVA"))){
       object$Ab <- object$Ab+CMSEPf(object)$Ab
       # object$Ab <- object$Ab+sdB(object)
       sdXcoef <- t(sqrt(apply(object$Ab,1,diag)))